chore: split dashboardView into two files
